requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
curl_cffi==0.13.0
filelock==3.20.3
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
narwhals==2.15.0
networkx==3.6.1
numpy==2.4.1
packaging==25.0
pandas==2.3.3
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.1
propcache==0.4.1
protobuf==6.33.3
psutil==7.2.1
pyarrow==22.0.0
pycparser==2.23
pydeck==0.9.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
torch-geometric==2.7.0
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
yfinance==1.0
